Moved adding header logic to LZ4Compression file
